refactor(volcengine): keep realtime helpers internal
